# [#24992] JFolder::delete : path isn't a folder administrator/components/com_kunena/archive , language...
git-svn-id: file:///Users/oratzes/svn/kunena@4462 daf2aa89-cb5b-0410-b8ee-fe93ebc9b27d

+++ branches/1.6-xillibit-19-02-2011/administrator/components/com_kunena/install/model.php
@@ -494,9 +494,9 @@ class KunenaModelInstall extends JModel {
 
 		// Cleanup directory structure
 		if (!Kunena::isSVN()) {
-			JFolder::delete(KPATH_ADMIN . '/archive');
-			JFolder::delete(KPATH_ADMIN . '/language');
-			JFolder::delete(KPATH_SITE . '/language');
+			if( JFolder::exists(KPATH_ADMIN . '/archive') ) JFolder::delete(KPATH_ADMIN . '/archive');
+			if( JFolder::exists(KPATH_ADMIN . '/language') ) JFolder::delete(KPATH_ADMIN . '/language');
+			if( JFolder::exists(KPATH_SITE . '/language') ) JFolder::delete(KPATH_SITE . '/language');
 		}
 
 		if (! $this->getError ()) {
